requirements.txt
Django==1.6
MySQL-python==1.2.4
argparse==1.2.1
django-debug-toolbar==0.11.0
django-redis-sessions==0.4.0
hiredis==0.1.1
python-memcached==1.53
redis==2.8.0
sqlparse==0.1.10
toredis==0.1.2
tornado==3.1.1
wsgiref==0.1.2
